Move subreddit name search to its own CF.
Subreddit name search is used for recommending
subreddits to users as they type a name on the
submission page.

Also adds a missing script that is needed
to load up the subreddit names into cassandra.
This script should be run occasionally (we do it
once per day) to fill/update the search cache.