Feature: Log out of other sessions
Currently the only way to log out of all sessions is to change your
password. Of course you can change your password to itself, however, this pull
request makes it a bit more explicit.

Underneath the covers it forces a change in the password salt by simply
changing the password to itself.